Revert "Review comments updated"
