+++ README.md
@@ -39,28 +39,14 @@ Before you start, ensure you have the following installed:
 Java 17
 Maven 3.6+
 A Relational Database (e.g., MySQL, PostgreSQL)
-### Configure Database Connection
-Configure your database connection in application.properties or application.yml. Here’s an example for application.properties:
-
-spring.datasource.url=jdbc:mysql://localhost:3306/your_database
-spring.datasource.username=your_username
-spring.datasource.password=your_password
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-
-### Initial Migration Scripts
-Created a directory src/main/resources/db/migration in project. Placed initial Flyway migration scripts in this directory. 
-The scripts names: V1__DB_IEMR.sql, V2__DB_IDENTITY.sql, V3__DB_REPORTING.sql.
 ### Creating Migrations
 Flyway migrations are SQL scripts located in the src/main/resources/db/migration directory. Each migration file should be named according to the Flyway convention: V1__Description.sql, V2__Description.sql, etc.
-
 ### Run Migrations
 * Give Database credentials in application.properties
 * Flyway automatically run migrations at application startup if you have configured it properly.
-
 ### Common Issues & Troubleshooting
 * Migration Failure : Check your SQL syntax and review the flyway_schema_history table for issues.
 * Database Connection Issues: Verify the connection details in application.properties.
-
 ### Contributing
 We welcome contributions! Please follow these steps:
