--- src/main/environment/common_ci.properties
@@ -1,23 +1,0 @@
-#db_iemr
-spring.datasource.dbiemr.url=@env.DATABASE_URL@
-spring.datasource.dbiemr.username=@env.DATABASE_USERNAME@
-spring.datasource.dbiemr.password=@env.DATABASE_PASSWORD@
-spring.datasource.dbiemr.driver-class-name=com.mysql.cj.jdbc.Driver
-
-#db_Identity
-spring.datasource.dbidentity.url=@env.IDENTITY_DATABASE_URL@
-spring.datasource.dbidentity.username=@env.IDENTITY_DATABASE_USERNAME@
-spring.datasource.dbidentity.password=@env.IDENTITY_DATABASE_PASSWORD@
-spring.datasource.dbidentity.driver-class-name=com.mysql.cj.jdbc.Driver
-
-#db_reporting
-spring.datasource.dbreporting.url=@env.REPORTING_DATABASE_URL@
-spring.datasource.dbreporting.username=@env.REPORTING_DATABASE_USERNAME@
-spring.datasource.dbreporting.password=@env.REPORTING_DATABASE_PASSWORD@
-spring.datasource.dbreporting.driver-class-name=com.mysql.cj.jdbc.Driver
-
-#1097_Identity
-spring.datasource.db1097identity.url=@env.1097_IDENTITY_DATABASE_URL@
-spring.datasource.db1097identity.username=@env.1097_IDENTITY_DATABASE_USERNAME@
-spring.datasource.db1097identity.password=@env.1097_IDENTITY_DATABASE_PASSWORD@
-spring.datasource.db1097identity.driver-class-name=com.mysql.cj.jdbc.Driver
