--- src/main/environment/common_local.properties
@@ -1,47 +1,0 @@
-spring.application.name=Amrit-DB
-
-spring.datasource.tomcat.initial-size=5
-spring.datasource.tomcat.max-active=30
-spring.datasource.tomcat.max-idle=15
-spring.datasource.tomcat.min-idle=5
-spring.datasource.tomcat.min-evictable-idle-time-millis=15000
-spring.datasource.tomcat.remove-abandoned=true
-spring.datasource.tomcat.logAbandoned=true
-spring.datasource.continue-on-error=true
-spring.datasource.tomcat.remove-abandoned-timeout=600
-spring.jpa.hibernate.naming.implicit-strategy=org.hibernate.boot.model.naming.ImplicitNamingStrategyLegacyJpaImpl
-spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
-spring.jpa.hibernate.naming_strategy=org.hibernate.cfg.EJB3NamingStrategy
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
-spring.jpa.hibernate.ddl-auto=none
-
-spring.datasource.dbiemr.url=<Enter AMRIT DB_IEMR URL here>
-spring.datasource.dbiemr.username=<Enter your AMRIT DB_IEMR username>
-spring.datasource.dbiemr.password=<Enter your AMRIT DB_IEMR password>
-spring.datasource.dbiemr.driver-class-name=com.mysql.cj.jdbc.Driver
-
-
-spring.datasource.dbidentity.url=<Enter AMRIT DB_IDENTITY URL here>
-spring.datasource.dbidentity.username=<Enter your AMRIT DB_IDENTITY username>
-spring.datasource.dbidentity.password=<Enter your AMRIT DB_IDENTITY password>
-spring.datasource.dbidentity.driver-class-name=com.mysql.cj.jdbc.Driver
-
-#db_reporting
-spring.datasource.dbreporting.url=<Enter AMRIT DB_REPORTING URL here>
-spring.datasource.dbreporting.username=<Enter your AMRIT DB_REPORTING username>
-spring.datasource.dbreporting.password=<Enter your AMRIT DB_REPORTING password>
-spring.datasource.dbreporting.driver-class-name=com.mysql.cj.jdbc.Driver
-
-spring.datasource.db1097identity.url=<Enter AMRIT DB_1097_IDENTITY URL here>
-spring.datasource.db1097identity.username=<Enter your AMRIT DB_1097_IDENTITY username>
-spring.datasource.db1097identity.password=<Enter your AMRIT DB_1097_IDENTITY password>
-spring.datasource.db1097identity.driver-class-name=com.mysql.cj.jdbc.Driver
-
-spring.flyway.locations=classpath:db/migration/dbiemr,classpath:db/migration/dbidentity,classpath:db/migration/dbreporting,classpath:db/migration/db1097identity
-spring.flyway.validateMigrationNaming=true
-spring.flyway.enabled=true
-spring.flyway.baseline-on-migrate=true
-spring.flyway.migrate=true
-
-logging.level.org.flywaydb=DEBUG
-logging.level.org.springframework=DEBUG
