+++ src/main/java/com/db/piramalswasthya/ServletInitializer.java
@@ -19,23 +19,16 @@
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see https://www.gnu.org/licenses/.
 */
-package com.piramalswasthya.db.data;
+package com.db.piramalswasthya;
 
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
 
-import java.sql.Timestamp;
+public class ServletInitializer extends SpringBootServletInitializer {
 
-import lombok.Data;
+	@Override
+	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+		return application.sources(AmritDbApplication.class);
+	}
 
-@Data
-public class FlywaySchemaVersion {
-	private int installedRank;
-	private String version;
-    private String description;
-    private String type;
-    private String script;
-    private Integer checksum;
-    private String installedBy;
-    private Timestamp installedOn;
-    private int executionTime;
-    private boolean success;
 }
