+++ src/main/java/com/db/piramalswasthya/controller/VersionController.java
@@ -19,7 +19,7 @@
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see https://www.gnu.org/licenses/.
 */
-package com.piramalswasthya.db.controller;
+package com.db.piramalswasthya.controller;
 
 import java.util.List;
 
@@ -29,33 +29,35 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.util.CollectionUtils;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.db.piramalswasthya.data.FlywaySchemaVersion;
+import com.db.piramalswasthya .service.VersionService;
 import com.google.gson.Gson;
-import com.piramalswasthya.db.data.FlywaySchemaVersion;
-import com.piramalswasthya.db.service.VersionService;
 
 @RestController
+
 @RequestMapping("/db/migration")
 public class VersionController {
 	private Logger logger = LoggerFactory.getLogger(VersionController.class);
-	@Autowired
+	
 	private VersionService service;
 	
+	@Autowired
+	public void setCommonServiceImpl(VersionService service) {
+		this.service = service;
+	}
+
 	@GetMapping("/version")
-	public String getLatestDBMigrationVersion(@RequestParam("database") String database) throws Exception {
+	public String getLatestDBMigrationVersion() throws Exception {
 		String resp = null;
 		try {
-			if(null!=database) {
-			List<FlywaySchemaVersion> latestVersion = service.getLatestVersion(database);
+			List<FlywaySchemaVersion> latestVersion = service.getLatestVersion();
 			Gson gson = new Gson();
 			if (null != latestVersion && !CollectionUtils.isEmpty(latestVersion)) {
 				resp = gson.toJson(latestVersion);
 			}
-			}
 		} catch (Exception e) {
-			logger.error("Error while getting DB Migration version {}: ", e.getMessage(), e);
 			throw new Exception("Error while getting DB Migration version: " + e.getMessage(), e);
 		}
 		return resp;
