+++ src/main/java/com/db/piramalswasthya/data/FlywaySchemaVersion.java
@@ -19,16 +19,23 @@
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see https://www.gnu.org/licenses/.
 */
-package com.piramalswasthya.db;
+package com.db.piramalswasthya.data;
 
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
 
-public class ServletInitializer extends SpringBootServletInitializer {
+import java.sql.Timestamp;
 
-	@Override
-	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(AmritDbApplication.class);
-	}
+import lombok.Data;
 
+@Data
+public class FlywaySchemaVersion {
+	private int installedRank;
+    private String version;
+    private String description;
+    private String type;
+    private String script;
+    private Integer checksum;
+    private String installedBy;
+    private Timestamp installedOn;
+    private int executionTime;
+    private boolean success;
 }
