+++ src/main/resources/application.properties
@@ -14,34 +14,13 @@ spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.Ph
 spring.jpa.hibernate.naming_strategy=org.hibernate.cfg.EJB3NamingStrategy
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
 spring.jpa.hibernate.ddl-auto=none
+spring.datasource.url=
+spring.datasource.username=
+spring.datasource.password=
+spring.datasource.driver-class-name=
 
-spring.datasource.dbiemr.url=<Enter AMRIT DB_IEMR URL here>
-spring.datasource.dbiemr.username=<Enter your AMRIT DB_IEMR username>
-spring.datasource.dbiemr.password=<Enter your AMRIT DB_IEMR password>
-spring.datasource.dbiemr.driver-class-name=com.mysql.cj.jdbc.Driver
-
-
-spring.datasource.dbidentity.url=<Enter AMRIT DB_IDENTITY URL here>
-spring.datasource.dbidentity.username=<Enter your AMRIT DB_IDENTITY username>
-spring.datasource.dbidentity.password=<Enter your AMRIT DB_IDENTITY password>
-spring.datasource.dbidentity.driver-class-name=com.mysql.cj.jdbc.Driver
-
-#db_reporting
-spring.datasource.dbreporting.url=<Enter AMRIT DB_REPORTING URL here>
-spring.datasource.dbreporting.username=<Enter your AMRIT DB_REPORTING username>
-spring.datasource.dbreporting.password=<Enter your AMRIT DB_REPORTING password>
-spring.datasource.dbreporting.driver-class-name=com.mysql.cj.jdbc.Driver
-
-spring.datasource.db1097identity.url=<Enter AMRIT DB_1097_IDENTITY URL here>
-spring.datasource.db1097identity.username=<Enter your AMRIT DB_1097_IDENTITY username>
-spring.datasource.db1097identity.password=<Enter your AMRIT DB_1097_IDENTITY password>
-spring.datasource.db1097identity.driver-class-name=com.mysql.cj.jdbc.Driver
-
-spring.flyway.locations=classpath:db/migration/dbiemr,classpath:db/migration/dbidentity,classpath:db/migration/dbreporting,classpath:db/migration/db1097identity
+spring.flyway.locations=classpath:db/migration
 spring.flyway.validateMigrationNaming=true
 spring.flyway.enabled=true
 spring.flyway.baseline-on-migrate=true
 spring.flyway.migrate=true
-
-logging.level.org.flywaydb=DEBUG
-logging.level.org.springframework=DEBUG
