+++ src/main/resources/db/migration/V2__Create_Table.sql
@@ -1,0 +1,9 @@
+CREATE TABLE db_iemr.usertoc (
+    id INT PRIMARY KEY AUTO_INCREMENT,
+    firstName VARCHAR(255) NOT NULL,
+    lastName VARCHAR(255) NOT NULL,
+    age INT NOT NULL,
+    salary DOUBLE NOT NULL,
+    companyName VARCHAR(255) NOT NULL,
+    isExist BOOLEAN NOT NULL
+);
