+++ src/test/java/com/db/piramalswasthya/AmritDbApplicationTests.java
@@ -1,0 +1,16 @@
+package com.db.piramalswasthya;
+
+import org.junit.jupiter.api.Test;
+import org.mockito.Mock;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AmritDbApplicationTests {
+	@Mock
+	AmritDbApplication amritDbApplication;
+
+	@Test
+	public void test() {
+		amritDbApplication.toString();
+	}
+}
